+++ examples/basic_app.js
@@ -5,7 +5,7 @@ var List = new PVCollection({
 	// collection options
 	name: 'Test list',
 	debug: true,
-	template: JST['test/templates/collection.hbs'],
+	template: JST['examples/templates/collection.hbs'],
 	functions: {
 		initialize: function(_options) {
 			this.log(['user init init'], 'info');
@@ -15,9 +15,9 @@ var List = new PVCollection({
 	// model options/template
 	debug: true,
 	foo: 'bar',
-	template: JST['test/templates/item.hbs'],
+	template: JST['examples/templates/item.hbs'],
 	functions: {
-		onchange: function(evt) {
+		onChange: function(evt) {
 			// this was added in model setup
 			this.render();
 		}